refactor(game): Lotto 클래스를 classicLotto 패키지로 이동
확장 가능한 구조를 위해 Lotto를 game/classicLotto로 이동
- Lotto 패키지 변경: lotto → lotto.game.classicLotto
- 모든 파일의 import 경로 수정
- LottoTest import 추가
